chore: Code style fixes
